critique: three small cleanups from cross-repo review
- Move calendarDot() from RemindersLib to GetClearKit; CalendarLib was
  defining the same function verbatim — now both pull from the shared source.
  Spec moves to GetClearKitTests accordingly.

- Move cleanLabel() from AppleContactKit boundary to ContactKit — pure
  string→string with no CN types, now testable at the right layer.
  CleanLabelSpec moves to ContactKitTests; AppleContactKit imports it unchanged.

- Split ReminderHandlersSpec (386-line monolith) into one spec file per
  handler. SpyStore moves to Fixtures.swift. No behavioral changes.
